Initial vermilion example content cleanup
Replace most of the existing example content with updated and simplified example
for running with Intel MPI.  Some notes about the changes:

- The original documentation suggested making a copy of the entire "examples"
directory structure.  Currently copies 122GB, when all that is needed to run the
example is a simple source file.  The instructions have been updated to show how
to copy just the necessary source file.
- The previous instructions for Intel MPI would not work for a multi-inode job.
It is necessary to add the IMPI_OFI_PROVIDER flag.
- The previous instructions for running with OpenMPI have several problems.
First, OpenMPI is not cimpiled with slurm support, so srun cannot be used.
Second, even with mpirun instead of srun, there are still errors when using
OpenMPI.  I am currently awaiting guidance for proper usage and will otherwise
update the docs to advise against using OpenMPI on multiple nodes.