Minor updates/fixes to Dask examples
Verified that dask-jobqueue runs on Kestrel.  With dask-mpi, i do get a
CommClosedError as mentioned in the Warning in the docs.  I have removed the
interface specifications, as ib0 doesn't exist on Kestrel.  The examples run
without the spec, but it may be good to verify what interface would be specified
on Kestrel (suspect it may be hsn0).